+++ leptos_workers/src/worker_message/mod.rs
@@ -1,0 +1,7 @@
+mod transferable;
+mod transferable_types;
+mod worker_message;
+
+pub use transferable::Transferable;
+
+pub(crate) use worker_message::{TransferableMessage, TransferableMessageType};
